ENH: refactor xarray (io/georef/vis)
