+++ wradlib/io/xarray.py
@@ -64,9 +64,8 @@ import datetime as dt
 import netCDF4 as nc
 import h5py
 import xarray as xr
-from osgeo import osr
 
-from ..georef import georeference_dataset, spherical_to_xyz, spherical_to_proj
+from ..georef import xarray
 
 
 # CfRadial 2.0 - ODIM_H5 mapping
@@ -317,156 +316,6 @@ global_variables = dict([('volume_number', np.int),
                          ('sweep_fixed_angle', (['sweep'], [np.nan])),
                          ('frequency', np.nan),
                          ('status_xml', 'None')])
-
-
-def as_xarray_dataarray(data, dims, coords):
-    """Create Xarray DataArray from NumPy Array
-
-        .. versionadded:: 1.3
-
-    Parameters
-    ----------
-    data : :class:`numpy:numpy.ndarray`
-        data array
-    dims : dictionary
-        dictionary describing xarray dimensions
-    coords : dictionary
-        dictionary describing xarray coordinates
-
-    Returns
-    -------
-    dataset : xr.DataArray
-        DataArray
-    """
-    da = xr.DataArray(data, coords=dims.values(), dims=dims.keys())
-    da = da.assign_coords(**coords)
-    return da
-
-
-def create_xarray_dataarray(data, r=None, phi=None, theta=None, proj=None,
-                            site=None, sweep_mode='azimuth_surveillance',
-                            rf=1.0, **kwargs):
-    """Create Xarray DataArray from Polar Radar Data
-
-        .. versionadded:: 1.3
-
-    Parameters
-    ----------
-    data : :class:`numpy:numpy.ndarray`
-        The data array. It is assumed that the first dimension is over
-        the azimuth angles, while the second dimension is over the range bins
-    r : :class:`numpy:numpy.ndarray`
-        The ranges. Units may be chosen arbitrarily, m preferred.
-    phi : :class:`numpy:numpy.ndarray`
-        The azimuth angles in degrees.
-    theta : :class:`numpy:numpy.ndarray`
-        The elevation angles in degrees.
-    proj : osr object
-        Destination Spatial Reference System (Projection).
-    site : tuple
-        Tuple of coordinates of the radar site.
-    sweep_mode : str
-        Defaults to 'PPI'.
-    rf : float
-        factor to scale range, defaults to 1. (no scale)
-
-    Keyword Arguments
-    -----------------
-    re : float
-        effective earth radius
-    ke : float
-        adjustment factor to account for the refractivity gradient that
-        affects radar beam propagation. Defaults to 4/3.
-    dim0 : str
-        Name of the first dimension. Defaults to 'azimuth'.
-    dim1 : str
-        Name of the second dimension. Defaults to 'range'.
-
-    Returns
-    -------
-    dataset : xr.DataArray
-        DataArray
-    """
-    if (r is None) or (phi is None) or (theta is None):
-        raise TypeError("wradlib: function `create_xarray_dataarray` requires "
-                        "r, phi and theta keyword-arguments.")
-
-    r = r.copy()
-    phi = phi.copy()
-    theta = theta.copy()
-
-    # create bins, rays 2D arrays for curvelinear coordinates
-    if sweep_mode == 'azimuth_surveillance':
-        bins, rays = np.meshgrid(r, phi, indexing='xy')
-    else:
-        bins, rays = np.meshgrid(r, theta, indexing='xy')
-
-    # setup for spherical earth calculations
-    re = kwargs.pop('re', None)
-    ke = kwargs.pop('ke', 4. / 3.)
-    if site is None:
-        site = (0., 0., 0.)
-        re = 6378137.
-
-    # GDAL OSR, convert to this proj
-    if isinstance(proj, osr.SpatialReference):
-        xyz = spherical_to_proj(r, phi, theta, site, proj=proj, re=re, ke=ke)
-        x = xyz[..., 0]
-        y = xyz[..., 1]
-        z = xyz[..., 2]
-    # other proj, convert to aeqd
-    elif proj:
-        xyz, proj = spherical_to_xyz(r, phi, theta, site, re=re, ke=ke,
-                                     squeeze=True)
-        x = xyz[..., 0]
-        y = xyz[..., 1]
-        z = xyz[..., 2]
-    # proj, convert to aeqd and add offset
-    else:
-        xyz, proj = spherical_to_xyz(r, phi, theta, site, re=re, ke=ke,
-                                     squeeze=True)
-        x = xyz[..., 0] + site[0]
-        y = xyz[..., 1] + site[1]
-        z = xyz[..., 2] + site[2]
-
-    # calculate center point
-    center = np.mean(xyz[:, 0, :], axis=0)
-
-    # calculate ground range
-    gr = np.sqrt((xyz[..., 0] - center[0]) ** 2 +
-                 (xyz[..., 1] - center[1]) ** 2)
-
-    # retrieve projection information
-    cs = []
-    if proj.IsProjected():
-        cs.append(proj.GetAttrValue('projcs'))
-    cs.append(proj.GetAttrValue('geogcs'))
-    projstr = ' - '.join(cs)
-
-    dims = collections.OrderedDict()
-    dim0 = kwargs.pop('dim0', 'azimuth')
-    dim1 = kwargs.pop('dim1', 'range')
-    dims[dim0] = np.arange(phi.shape[0])
-    dims[dim1] = r / rf
-    coords = {'azimuth': ([dim0], phi),
-              'elevation': ([dim0], theta),
-              'bins': ([dim0, dim1], bins / rf),
-              'rays': ([dim0, dim1], rays),
-              'x': ([dim0, dim1], x / rf),
-              'y': ([dim0, dim1], y / rf),
-              'z': ([dim0, dim1], z / rf),
-              'gr': ([dim0, dim1], gr / rf),
-              'longitude': (site[0]),
-              'latitude': (site[1]),
-              'altitude': (site[2]),
-              'sweep_mode': sweep_mode,
-              'projection': projstr,
-              }
-
-    # create xarray dataarray
-    da = as_xarray_dataarray(data, dims=dims, coords=coords)
-
-    return da
 
 
 @xr.register_dataset_accessor('gamic')
@@ -1004,7 +853,7 @@ class XRadVol(collections.abc.MutableMapping):
 
     def georeference(self):
         for swp in self:
-            self[swp] = self[swp].pipe(georeference_dataset)
+            self[swp] = self[swp].pipe(xarray.georeference_dataset)
 
 
 class CfRadial(XRadVol):
@@ -1050,7 +899,7 @@ class CfRadial(XRadVol):
 
             # adding xyz aeqd-coordinates
             if georef:
-                ds = georeference_dataset(ds)
+                ds = xarray.georeference_dataset(ds)
 
             self._sweeps[sw] = ds
 
@@ -1098,7 +947,7 @@ class CfRadial(XRadVol):
 
             # adding xyz aeqd-coordinates
             if georef:
-                ds = georeference_dataset(ds)
+                ds = xarray.georeference_dataset(ds)
 
             self._sweeps[sw] = ds
 
@@ -1273,7 +1122,7 @@ class OdimH5(XRadVol):
                 # georeference needs coordinate variables
                 if georef:
                     geods = xr.Dataset(vars, coords)
-                    geods = georeference_dataset(geods)
+                    geods = xarray.georeference_dataset(geods)
                     coords.update(geods.coords)
 
             # time coordinate
